confirm

+++ admin/view/cate/cate.php
@@ -35,7 +35,7 @@
                     </td>
     	            	<td>
     	            		<a href="#" class="save"><i class="fa fa-edit"></i>Lưu lại</a> | 
-                      <a href="?action=deleteCate&id=<?php echo $value['category_id']; ?>"><i class="fa fa-trash"></i>Xóa</a>
+                      <a onclick="return confirm('Bạn có chắc muốn xóa')" href="?action=deleteCate&id=<?php echo $value['category_id']; ?>"><i class="fa fa-trash"></i>Xóa</a>
     	            		<div class="message"></div>
     	            	</td>	
     	              </tr>
